lang/eval: Deciding a provider instance for each resource instance
This introduces the machinery for representing provider configs, provider
instances, and the relationships between resource instances and provider
instances.

Unfortunately there's a lot of accumulated complexity in the rules for
how OpenTofu currently decides which provider configs are available for
use in a module, including lots of implicit behavior for
backward-compatibility, so right now that is stubbed out with just a
hard-coded implementation that puts a "test/foo" provider with local name
"foo" in every module instance. We'll deal with a real implementation of
all that in later commits.

Signed-off-by: Martin Atkins <mart@degeneration.co.uk>